fix: [readequação] correção ao finalizar readequação de plano de divulgação
Former-commit-id: 6de27757b3432c942f6738582fe0fdac4ef9cbb4

+++ application/modules/default/controllers/ReadequacoesController.php
@@ -3777,7 +3777,9 @@ class ReadequacoesController extends GenericControllerNew {
                             $PlanoDivulgacaoEmQuestao = $PlanoDeDivulgacao->buscar(array('idProjeto = ?'=>$dadosPrj->idProjeto, 'idPeca = ?'=>$plano->idPeca, 'idVeiculo = ?'=>$plano->idVeiculo))->current();
                             $tbLogomarca = new tbLogomarca();
                             $dadosLogomarcaDaDivulgacao = $tbLogomarca->buscar(array('idPlanoDivulgacao = ?' => $PlanoDivulgacaoEmQuestao->idPlanoDivulgacao))->current();
-                            $dadosLogomarcaDaDivulgacao->delete();
+                            if (!empty($dadosLogomarcaDaDivulgacao)) {
+                                $dadosLogomarcaDaDivulgacao->delete();
+                            }
                             $PlanoDivulgacaoEmQuestao->delete();
 
                         } else if($plano->tpSolicitacao == 'I') { //Se o plano de divulgação foi incluído, cria um novo registro na tabela SAC.dbo.PlanoDeDivulgacao
